gcs-to-pubsub-publisher/requirements.txt
# Function dependencies, for example:
# package>=version

google-cloud-storage>=2.5.0
google-auth-oauthlib>=0.5.2
pandas>=1.4.3